chore: handle compressed files
